chore(lint): pin ruff 0.16, select rules explicitly, fix new findings
A fresh `pip install -e ".[dev]"` resolved `ruff>=0.4` to 0.16.7,
whose built-in default rule set (unselected) is far broader than
0.14.11's (verified empirically with `ruff check --isolated`: 0.14's
default is E4/E7/E9/F only, 0.16's spans ~35 plugin categories). That
left what gets checked dependent on whichever ruff happened to be
installed. Fix: `ruff check --fix .` plus minimal manual edits for
everything not auto-fixable, then pin `[tool.ruff.lint] select`
explicitly to the old baseline (E4/E7/E9/F) plus the exact new rule
codes 0.16 surfaced on this codebase (B017, BLE001, I001, PIE790,
PLR0402, RUF007, RUF022, RUF046, SIM102, UP007, UP024, UP035, W605),
so upgrading ruff again won't silently expand what's enforced. Also
set `isort.combine-as-imports = true` so sorting `as`-aliased
re-exports (e.g. src/qpl/pricing.py) doesn't split them across
multiple import statements.

66 findings fixed: 46 auto-fixed (import sorting, __all__ sorting,
os-error-alias, deprecated-import, invalid-escape-sequence,
manual-from-import, unnecessary-placeholder), 20 by hand:
- zip(seq, seq[1:]) -> itertools.pairwise(seq) (RUF007, 5x).
- int(round(...)) -> round(...), already int (RUF046, 1x).
- Union[float, int, np.ndarray] -> float | int | np.ndarray (UP007, 2x).
- nested if collapsed with `and` (SIM102, 1x), no behavior change.
- pytest.raises(Exception) -> pytest.raises(dataclasses.FrozenInstanceError)
  for a frozen-dataclass mutation test (B017, 1x); FrozenInstanceError
  is an AttributeError subclass, so this is a narrowing, not a broadening.
- blind `except Exception` (BLE001, 10x): narrowed to the callee's
  actual documented exception types where the callee only raises
  specific ones (implied_vol_demo.py, the notebook cell inverting IV,
  historical_vol_demo.py's log_returns call); left as `except Exception`
  with a `# noqa: BLE001` justification comment where the catch is a
  deliberate translate-any-failure boundary (the implied_volatility
  solver wrapping brentq, market/data.py's cache-read and fetch paths,
  and the remaining fetch-fallback demo/notebook cells) rather than a
  narrowing that could otherwise silently let an unwrapped exception
  type escape a documented contract.

No behavior changes. `ruff check .` and `pytest -q` (171 passed) are
green.

+++ src/qpl/engines/pde/pricers.py
@@ -189,7 +189,7 @@ def price_european(
 
     if cfg.strike_alignment == "midpoint":
         # Nearest half-integer node position for the strike; see the docstring.
-        j = max(int(round(k / ds - 0.5)), 0)
+        j = max(round(k / ds - 0.5), 0)
         ds = k / (j + 0.5)
         s_max = ds * n_s
